Partially support building on Windows (64-bit)
jitana-graph and test cases can now be compiled on Windows. The
following configuration was used to test:

- Windows 10 running on VMWare Fusion 8
- Visual Studio 2015
- CMake 3.4.0
- Boost 1.59.0

jitana-travis cannot be compiled because it relies on POSIX.

+++ include/jitana/vm_core/class_loader.hpp
@@ -106,7 +106,7 @@ namespace jitana {
                 file_hdl.loader_hdl = hdl;
                 std::for_each(filenames_first, filenames_last,
                               [&](const auto& s) {
-                                  file_hdl.idx = dex_files.size();
+                                  file_hdl.idx = uint8_t(dex_files.size());
                                   dex_files.emplace_back(file_hdl, s);
                               });
             }
